requirements.txt
pandas
numpy
scikit-learn
matplotlib
seaborn
python-pptx
joblib
pyinstaller
chardet
